requirements.txt
aiohttp==3.6.2
pycrypto==2.6.1
request==2.22.0
xlsxwriter==1.2.2